Makes larvae and facehuggers able to use custom emotes (#5496)
# About the pull request

Makes larvae and facehuggers able to use custom emotes with the 'me'
verb. (All other emotes still work the same as before.)

(The code just being copy+pasted between the two castes isn't great, but
I can't really think of a cleaner way of doing it.)

# Explain why it's good for the game

Larva RP is one of the most vital parts of the game.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Made larvae and facehuggers able to use custom emotes.
/:cl: