Ap rocket stun and damage consistency and renames vars in rocket.dm (#5183)
# About the pull request

renames one latter vars in rocket.dm. Makes AP rocket behavior more
consistent, AP rocked stuns you for TWICE as longs when it lands on turf
under the mob and does not hit it directly. to compensate that I
increased the stun duration by 50% to avrage of the turf hit and mob
hit. the PR also reduces max distance damage, as it was higer for no
aperent reason
but I will not fight agains it if maintainer decides that the values
should be diferent, I do not care.
# Explain why it's good for the game
expecting 2,6 sec stun and geting 5.3 when someone knows to aim for the
tile under you and not you is not fun.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: AP rocket applies 3 weaken and 3 paralyze no matter if you hit
the mob directly(from 2), at max range (from 2) or tile under it (from
4)
balance: indirect AP rocket does the same damage to walls and mobs
/:cl: