Cleans up defib code a little (#5507)
# About the pull request
removes the "heart_already_damaged" var from the defib file, instead
just having the defib check if it failed before applying the random
chance for heart damage
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

![image](https://github.com/cmss13-devs/cmss13/assets/54422837/631fe43d-d77a-4d3c-b3ff-1a422298dd50)
# Testing Photographs and Procedure
<details>
<summary>Test here</summary>


https://github.com/cmss13-devs/cmss13/assets/54422837/cd5753fb-0909-4957-8867-a06a093eaacc


Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
no player facing changes