Finishes #5128 (Paygrade defines) & Fixes Syndie ID (#5260)
# About the pull request
As title. This PR continues my work from #5128 and adds the defines for
PMCs, Mercenaries, VAI and TWE.
I have also changed how the global lists for highcommand (both USCM and
WY) work and have made a new global list of all USCM officers to allow
for easy reference in #5202.
To do the above I have added a default_faction var on paygrades, along
with a variable to indicate if a paygrade refers to an "officer" or not,
and if the officer is a "flag" officer, for senior leadership.
I have also updated the Syndicate ID card to work properly.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Adds paygrade defines for PMCs, VAI, Mercenaries and TWE.
code: Adds code to populate reference lists for certain paygrades.
fix: Fixes PMC synth not using the correct grade.
add: Syndicate ID card repathed and works with paygrades.
/:cl: