Fixes a bunch of spelling mistakes (#5555)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->
Most of the changes here are minor, mostly to comments or descriptions.
Here is the full list.

- recieve -> receive
- percieve -> perceive
- alot -> a lot
- atleast -> at least
- succesfully -> successfully
- seperate -> separate
- noticable -> noticeable
- concious -> conscious
- enviroment -> environment
- existance -> existence
- foward -> forward
- interupted -> interrupted
- independant -> independent
- guarenteed -> guaranteed
- overriden -> overridden
- occuring -> occurring
- occured -> occurred
- transciever -> transceiver
- maintenace -> maintenance

# Explain why it's good for the game
Makes it easier to read things

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Code compiles and runs

</details>


# Changelog
:cl:
spellcheck: Fixed a few uncommon typos across the entire codebase
/:cl: