Tweak Game Mode Start Handling (#5506)
# About the pull request

Whisky Outpost is so rare now that it has taken two months for anyone to
notice that it has been throwing a runtime when determining how many
players are required for the game mode (140). Additionally, the Ticker
subsystem really poorly handled this situation.

This PR makes it so:
- Admins can bypass checks to start a game mode
- Messages when a game mode fails to start is more verbose
- A game mode that fails to start will attempt again once
- Otherwise it will set game mode to default and initiate a game map
vote and restart if there are no active admins; or it will delay start
and message admins.

# Explain why it's good for the game

This should hopefully be less confusing to people what's going on when a
game mode like WO fails to start and give admins an easier time to
handle situations like this.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>
No Admins (or afk admins):


![no_admins](https://github.com/cmss13-devs/cmss13/assets/76988376/8491981a-e740-4556-ac86-dd038d31c894)

With Admins:


![admins](https://github.com/cmss13-devs/cmss13/assets/76988376/0f0652a9-bd72-4993-a568-e3aed1ac6e6d)

</details>


# Changelog
:cl: Drathek
add: The game will now change game mode to default and initiate a map
vote if a game mode fails to start twice and there are no admins online.
fix: Fixed a runtime during game mode voting
admin: Admins can now use the Start Round verb to bypass a gamemode's
checks
/:cl: