Adds a Researcher and CMB variants on Trijent, also touched up RD Clothes and the Trijent Sec Guard. (#5438)
# About the pull request

Don't worry Synth mains, the **/obj/item/clothing/under/rank/rdalt**
path only leads to the RD's Locker, so I didn't remove your uniform.

Added a second RD uniform, which IMO has a better sprite than the
current one I changed.

Also added a Researcher and CMB Variant on Trijent, along with changing
the current Sec Guard as he was using USCM MP Gear, now he's using
proper Wey-Yu gear.

# Explain why it's good for the game

Surprised Trijent didn't have a Researcher or a special CMB spawn like
the rest of the maps, no reason as to why it shouldn't.. it has a lab
afterall.

The Trijent Security Guard had an MP hat, I don't think he's apart of
the USCM MPs so I removed that and changed it to a nice looking PMC hat,
along with a Wey-Yu Security Vest. Truly a Company moment.

Added an RD outfit for fluff for the new Researcher role and replaced
the unused synthetic service uniform with it's RD counterpart as well,
which could be used in the future.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added a Researcher on Trijent.
add: Added a CMB Deputy on Trijent.
add: Added a second RD uniform.
code: changed the Trijent Security Guard to not use a USCM MP Hat and
trenchcoat.
code: changed a not used "synthetic service uniform" to it's before
shamelessly taken RD Variant.
/:cl:

---------

Co-authored-by: Jeff Watchson <johncenatoutourourou@gmail.com>