UPP ERT Backstory Change (#5559)
# About the pull request

The UPP ERT provides a backstory that mentions they're here to strike at
the USMC even if friendly, which is at odds with the objectives a
friendly UPP ERT will receive. This is because friendly UPP are a
comparatively recent change and the backstory was never tweaked to
account for it. This PR makes it so that if the ERT is friendly, the
final line is changed, otherwise it'll be the same. I've also added some
slight variability to the synthetic background so they can roleplay
around that if they wish.

I tested the changes and it appears to work properly. I can't find any
issue reports about this, otherwise I would have linked them.

# Explain why it's good for the game

Reduces confusion on the part of players who read the big text and not
the smaller text.


# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>

The code runs.

</details>


# Changelog
:cl:
add: Added a new backstory blurb for friendly UPP and some variance to
the backstory of UPP synthetics.
/:cl:

---------

Co-authored-by: WickedCybs <wickedcybs@gmail.com>