Chestbursting larva no longer get hive orders when there are no hive orders (#5576)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

If there are no hiveorders, larva that chest burst no longer are told
that their orders are to "".

# Explain why it's good for the game
Its not very helpful when theres no orders.

![image](https://github.com/cmss13-devs/cmss13/assets/91113370/1170f802-f66d-4745-ad16-4300181d017c)



# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/91113370/dfc90648-aefa-42a8-a71c-fc9ed064a541)

</details>


# Changelog
:cl:
fix: Chestbursting larva no longer get hive orders when there are no
hive orders
/:cl: