Réorganisation de l'équipe médicale - 1 (#5458)
# About the pull request

1. Re-organizes which is the gear presented for medical in their preps,
such as fixing wrong code within the CMO preps that didn't specify the
combat equipment as combat equipment in the code, consequently making
the player need to decide between a labcoat, or the damn combat armor.
Adds a syringe case to all 'essentials' packs, and gives nurses a basic
surgical kit to support doctors, in the same way combat medics have.
Adds a small variation of the classic chef's apron caled 'Medical's
apron', being available for all medical personnel but being the only
'suit' option for nurses, to make them more reconnaissable between the
medical team. Separates their snow gear from the usual gear, so now they
can take their suit AND a snowcoat, instead of one or another, also
gives them a balaclava and snow scarf options for the cold weather. Adds
standard backpacks and satchels to all medical's preps. Personnel will
now be able to choose between the normal HUD or the prescripted one in
their preps.
2. All personnel will now wake up with less equipment than before, with
the CMO being stripped from all their equipment in the same way the rest
of the medical, waking up only with their uniform, shoes and obviously a
backpack. CMOs and Researchers don't wake up with a labcoat anymore, but
their research papers and...syringe, now will wake up in their pants'
pockets (this alone will probably help people to not forget they have
that stuff in the damn labcoat pocket). They will not wake up with HUDs
anymore, but the Prescription option that would wake up with them is now
available on their usual preps. Given the CMO's now 'lackage' of awaking
equipment, their vendor preps have been completely updated to the modern
parameters and now features everything (and a little more) that they
would wake up with, including the afore mentioned 'essential' packs,
that have been designed by our comrade Julian a while ago.
3. Adds a specific latejoin for medical, so now they will wake up in the
same spot as the firstjoiners, making their walk between cryo and their
workplace significantly smaller.

# Explain why it's good for the game

Its supposed to give a bigger Quality of Life for players and to
approach the roles to a more coherent line with the other roles, such as
the normal marines, POs, and the Engineering team.

# Testing Photographs and Procedure
N/A


</details>


# Changelog

:cl:
add: Adds a variation of the chef's apron, "Medical's apron", to be used
by medical personnel, mainly nurses.
fix: Fixed the code in the prep vendors to make things of different
groups not act like they are of the same group, like how you couldn't
take a labcoat and a snowcoat, only one of both, which does not make
sense.
code: Update of the medical's vendor preps. Exclusion of most part of
medical's equipment when waking up from cryo. Addition of latejoin
landmarks to the same place where they usually wake up as firstjoiners.
maptweak: Added the landmarks to the map and changed the CMO's vendor to
an unused space in his room to improve on item grabbing, because the
items were being covered by the bed's sheet, and making player's life to
take things from it more difficult. Also re-added a bedsheet in the
medical's area that some old PR somehow removed.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>