Green Beret sprite fix (#5566)
# About the pull request

fixes #3573
Reuses existing rmc beret, which are green. it fits the beret theme, and
compared to red beret, it is a great match.

Please do note that black and red beret have weird sprites compared to
other beret. This is not the case when worn, however something to note.

# Explain why it's good for the game

Invisible sprites are bad
# Testing Photographs and Procedure

--Testing--

- Local Server
- Character Pref
- USMC beret, green
- spawn

<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/66401072/499032ca-822c-428d-9a20-bfce149f1e9b)


![image](https://github.com/cmss13-devs/cmss13/assets/66401072/09eb4694-6fdc-4c20-a042-47b608a72f79)


![image](https://github.com/cmss13-devs/cmss13/assets/66401072/d5ab6375-7a87-4798-8109-c0263bfc00aa)


![image](https://github.com/cmss13-devs/cmss13/assets/66401072/a5db5e77-aac3-4a51-afa4-91579db0bd96)

</details>


# Changelog
:cl:
fix: character pref Green Beret is no longer invisible
/:cl: