Added a USCMC Recruiter Survivor on Solaris. (#5474)
# About the pull request

Adds a USCMC Recruiter survivor to Solaris and a USCMC Survivor template
to the code

Changed the Solaris 'Ridge' CL and fluffed up the ARES announcement text
when everyone wakes up

# Explain why it's good for the game

Well, everything being standardised is good, due to every other surv
having no 'Ridge' next to Solaris in their assignments, unlike the CL
which he does, so this changes it to make him fit with everyone else.
(Or the opposite can be done, giving everyone else Ridge next to
Solaris. Whatever maintainers want.)

More survivor roles which are different and unique I hope are greatly
appriciated and I saw an opportunity due to the map having some sort of
USCMC insert, I think it'll add more Roleplay, which is never a bad
thing.

Also a template to have and to use incase there are any more USCMC
future survivors. (Currently using the security skills, perhaps this
should change in the future.)


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added a USCMC Recruiter to the Solaris surv pool.
add: Added a USCM Survivor template anycase anyone wants to add USCM
Survivors to other maps.
fix: Every other surv has no 'Ridge' next to Solaris in their
assignments, only the CL. Fixes this.
spellcheck: Embellished the Solaris Ridge announce text, to make it so
the Company and more fluff about it's actual role like the wiki states.
/:cl:

---------

Co-authored-by: Jeff Watchson <johncenatoutourourou@gmail.com>