Bump alternative tests to 515.1630 (#5632)
# About the pull request

This PR bumps the 515 alternative test to 515.1630 in github checks now
that we have updated the server to that version. It also caused a couple
more hard deletes; one of which (for globals) is sort of a byond issue
but TG had a workaround for it.

# Explain why it's good for the game

Tests should replicate what is used for live.

# Testing Photographs and Procedure

See Checks.

# Changelog

No player facing changes.