Revert damage changes in "MP QoL: part 1" (#5647)
# About the pull request

This PR reverts the damage changes in #5554 

# Explain why it's good for the game

Batons don't need to be more deadly than they were. Tazers do not need
to cause blood effects because of brute damage applied.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Drathek
balance: Reverted damage changes to tazers and batons.
/:cl: