Moves a table, adds a photocopier to brig. (#5622)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
--> Adds a Photocopier to the Brig. Moves a table and it's contents out
the way a little. Specifically the camera and it's films two tiles north
of where they previously were.

# Explain why it's good for the game
Allows MPs to photocopy photos and papers without barging into CIC or
going elsewhere. A small QOL tweak.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![EveryoneLovesPaperwork](https://github.com/cmss13-devs/cmss13/assets/7815982/b9044efb-c3aa-409c-a0af-deca1a36db1c)


![EveryoneLovesPaperwork2](https://github.com/cmss13-devs/cmss13/assets/7815982/a18989b2-b702-43a5-a0f0-448d69f04ea8)


Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
maptweak: Add Brig Photocopier
/:cl: