Helmet HUDs skip over HUDs that are already in effect on the player (#5600)
# About the pull request

Makes it so that if you have an active HUD, a helmet visor that does the
same thing will not be considered when switching through visors in your
helmet. Credit to harryob for his help with the code.

# Explain why it's good for the game

Visors are usually a pain to switch through when their functionality
isn't needed, take for example, the squad optic, which does the same job
as the headset (HUD-wise). This change makes it so you don't need to
switch through unneeded HUDs.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Stakeyng
qol: Removes the need to cycle through helmet visors that contain HUD
effects that are already active on your character.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>