Revert "refactor : turn some "# and use the proper define from colour.dm for 21 files." (#5626)
Reverts cmss13-devs/cmss13#5553
# About the pull request
Revert because some files are case sensitive and i did those change on a
bunch of them so better remove it all.
it wasn't an important PR anyway...
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
# Testing Photographs and Procedure
i tested welding different different cades at the same time seem to work
i had some weirdness but it's working.
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: revert PR #5553 as it is causing issues with color on atmo pipe.
/:cl: