Removes the two gauss sentry guns someone accidently boxxed into ice 3 over a year ago that has somehow survived multiple bugfixes and a rework of the map (#5648)
# About the pull request

title

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

bug


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: SpartanBobby
maptweak: Fixes the two gauss sentry guns that were boxxed in by a
mapper over a year ago on ice 3
/:cl: