added squads marine ability to buy more sidearm magazines and moved/changed some section in vendors around. (#5605)
# About the pull request
1-Main idea is to give players more option when they are equipping...
2-while i was at it i moved some section in team leader vendor to be
more in line with other squad marine vendors...

3-concerning the price i fallowed what was already in place.
i preserved the old imbalance that is between different roles concerning
ammo prices...
took as a model the intelligence officer vendor concerning prices of new
magazine etc...
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
More option is always good.
my main concern would be balance but since those magazine are available
freely in limited quantity i don't think it would cause an issue...

i think it also will help players to know where things are...

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: added two new section to replace ammunition section primary
ammunition and sidearm ammunition and rearranged it for team leader.
balance: added the ability for comtech, medic, and squad leader to buy
M4A3 (HP and AP) and VP78 magazines for 3 points and M44 Heavy for 6
points.
balance: added the ability for SG to buy M4A3 (HP and AP) and VP78
magazines for 5 points and M44 Heavy for 10 points.
balance: added the ability for rifleman to buy M4A3 (HP and AP) and VP78
magazines for 5 points.
/:cl:

Co-authored-by: Julien <jverger.ingx@gmail.com>